edgeManagement/backend/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.1
Flask==1.1.1
flask-restx==0.2.0
flask-cors==3.0.10
gunicorn==20.0.4
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kubernetes==19.15.0
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.6
pyrsistent==0.16.0
pytest==5.4.1
python-decouple==3.4
pytz==2019.3
six==1.14.0
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
